feat: Calculate and save the average Age of Information (AoI)
This commit adds the calculation and saving of the average Age of Information (AoI) for each run. The average AoI is computed by summing up the AoI values for each ASN and dividing it by the total number of ASNs. The result is then multiplied by the slot duration to get the average AoI in seconds. This information is added to the global stats of each run.